DeliveryFood/requirements.txt
Django==5.1.3
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
drf-yasg==1.21.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-simple-history==3.7.0
django-filter==24.3
django-import-export==4.3.3
openpyxl==3.1.5
pylint==3.3.3
pylint-django==2.6.1
celery==5.4.0